Server/requirements.txt
flask
flask-cors
flask-socketio
tensorflow
numpy
Pillow
